refactor: split the modules this round grew past the readability limit
The token-signal contract and the ATIF mapping both grew their test modules
past the 500-line production-file gate; both move to sibling `*_test.rs`
files, which is how the rest of the tree keeps large suites out of the
production count.

`conformance/walk.rs` splits along a real seam rather than a line count: how
a spec is read — an action's legal source states, which states end the run,
which actions are emissions rather than transitions — is now
`conformance/spec_view.rs`, and the walk asks it instead of interpreting the
spec itself.
